requirements.txt
asgiref==3.6.0
blinker==1.6.2
camelot-py==0.9.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
distlib==0.3.6
distro==1.8.0
Django==4.2.1
docx2pdf==0.1.8
et-xmlfile==1.1.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
fonttools==4.39.4
fpdf==1.7.2
gevent==22.10.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.1
pdf2docx==0.5.6
pdf2image==1.16.3
pdf2xlsx==1.2.0
pdfkit==1.0.0
pdfminer.six==20221105
pdftabextract==0.3.0
Pillow==9.5.0
platformdirs==3.5.1
pycparser==2.21
PyExcelerate==0.10.0
PyMuPDF==1.22.3
PyPDF2==2.12.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pywin32==306
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tabula==1.0.5
tabula-py==2.7.0
termcolor==2.3.0
tqdm==4.65.0
tzdata==2023.3
virtualenv==20.23.0
waitress==2.1.2
Werkzeug==2.3.4
zope.event==4.6
zope.interface==6.0
